Value parsing moved to a separate module
